requirements.txt
mkdocs
mkdocs-material
mkdocs-gen-files
mkdocs-macros-plugin
pyyaml